fix: rename consumer
