+++ src/Consumer/Nova/Helper.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Taurus\Workflow\Consumer\Nova;
+
+class Helper {}
