--- src/Consumer/OdysseyNova/Helper.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Taurus\Workflow\Consumer\OdysseyNova;
-
-use Carbon\Carbon;
-use Illuminate\Support\Facades\Config;
-use Illuminate\Support\Facades\Storage;
-
-class Helper {}
